@spaolacci 2019-03-17
Sorry for the more than late answer. I can't observe the allocation anymore as of Go 1.12 at least - I guess escape analysis eventually found it somewhere in the between - but those two references are intentionally useless, so better without them.

I did run a more comprehensive allocation check. All the regular calls are zero allocation, streaming hashers are exactly one allocation.